만분율 확장함수 추가

+++ Consts/GameRuleConst.cs
@@ -1,4 +1,4 @@
 internal static class GameRuleConst
 {
-    public const float PERCENTAGE_BASE = 10000f;
+    public const int PERCENTAGE_BASE = 10000;
 }
